requirements.txt
azure-functions==1.16.0
azure-identity==1.15.0
azure-storage-blob==12.14.1
pandas==1.5.3
psutil==5.9.5
openpyxl==3.1.2